+++ src/main/java/io/appulse/utils/BytesByteBuffer.java
@@ -47,7 +47,7 @@ class BytesByteBuffer extends BytesAbstract {
   }
 
   @Override
-  public boolean isResizable () {
+  public boolean isAutoResizable () {
     return false;
   }